# src/main/java/com/mahindra/api/model/Country.java
package com.mahindra.api.model;

import jakarta.persistence.*;

@Entity
@Table(name = "countries",
        uniqueConstraints = @UniqueConstraint(name = "uq_country_cca2", columnNames = "cca2"))
public class Country {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String name;
    private String capital;
    private Long population;
    private String region;
    private Double area;

    @Column(name = "flag_url")
    private String flagUrl;

    private String cca2;

    @Column(name = "phone_code")
    private String phoneCode;

    private String emoji;

    @Column(name = "capital_lat")
    private Double capitalLat;

    @Column(name = "capital_lng")
    private Double capitalLng;

    /** Required by JPA. */
    public Country() {}

    /** Convenience constructor for tests with minimal data. */
    public Country(Long id, String name) {
        this.id = id;
        this.name = name;
    }

    /** Full constructor for programmatic creation (ID generated by JPA). */
    public Country(String name, String capital, Long population, String region,
                   Double area, String flagUrl, String cca2) {
        this.name = name;
        this.capital = capital;
        this.population = population;
        this.region = region;
        this.area = area;
        this.flagUrl = flagUrl;
        this.cca2 = cca2;
    }

    public Long getId()          { return id; }
    public void setId(Long id)   { this.id = id; }

    public String getName()             { return name; }
    public void setName(String name)    { this.name = name; }

    public String getCapital()              { return capital; }
    public void setCapital(String capital)  { this.capital = capital; }

    public Long getPopulation()                 { return population; }
    public void setPopulation(Long population)  { this.population = population; }

    public String getRegion()               { return region; }
    public void setRegion(String region)    { this.region = region; }

    public Double getArea()             { return area; }
    public void setArea(Double area)    { this.area = area; }

    public String getFlagUrl()              { return flagUrl; }
    public void setFlagUrl(String flagUrl)  { this.flagUrl = flagUrl; }

    public String getCca2()             { return cca2; }
    public void setCca2(String cca2)    { this.cca2 = cca2; }

    public String getPhoneCode()                  { return phoneCode; }
    public void setPhoneCode(String phoneCode)    { this.phoneCode = phoneCode; }

    public String getEmoji()              { return emoji; }
    public void setEmoji(String emoji)    { this.emoji = emoji; }

    public Double getCapitalLat()                   { return capitalLat; }
    public void setCapitalLat(Double capitalLat)    { this.capitalLat = capitalLat; }

    public Double getCapitalLng()                   { return capitalLng; }
    public void setCapitalLng(Double capitalLng)    { this.capitalLng = capitalLng; }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (!(o instanceof Country other)) return false;
        return id != null && id.equals(other.id);
    }

    @Override
    public int hashCode() { return getClass().hashCode(); }
}