# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Mahindra Geo API
  version: 1.0.0
  description: |
    REST API for browsing countries and their cities.
    Supports paginated city listings and city detail lookups.

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Countries
    description: Operations related to countries
  - name: Cities
    description: Operations related to cities

paths:
  /countries:
    get:
      operationId: listCountries
      summary: List all countries
      description: Returns a complete list of all available countries.
      tags:
        - Countries
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
              example:
                - id: 1
                  name: India
                - id: 2
                  name: United States

  /countries/{countryId}/cities:
    get:
      operationId: listCitiesByCountry
      summary: List cities for a country (paginated)
      description: Returns a paginated list of cities belonging to the specified country.
      tags:
        - Cities
      parameters:
        - name: countryId
          in: path
          required: true
          description: The ID of the country
          schema:
            type: integer
            format: int64
            example: 1
        - name: page
          in: query
          required: false
          description: Zero-based page index (0..N)
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: size
          in: query
          required: false
          description: Number of records per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCityResponse'
        '404':
          description: Country not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cities/{cityId}:
    get:
      operationId: getCityById
      summary: Get city details by ID
      description: Returns detailed information for a specific city.
      tags:
        - Cities
      parameters:
        - name: cityId
          in: path
          required: true
          description: The ID of the city
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: City details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '404':
          description: City not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Country:
      type: object
      description: Represents a country
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the country
          example: 1
        name:
          type: string
          description: Name of the country
          example: India

    City:
      type: object
      description: Represents a city with optional details
      required:
        - id
        - name
        - countryId
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the city
          example: 1
        name:
          type: string
          description: Name of the city
          example: Mumbai
        countryId:
          type: integer
          format: int64
          description: ID of the country this city belongs to
          example: 1
        population:
          type: integer
          format: int64
          description: Population of the city (optional)
          example: 20667656
          nullable: true
        zipCode:
          type: string
          description: Primary ZIP/postal code (optional)
          example: "400001"
          nullable: true
        description:
          type: string
          description: Short description of the city (optional)
          example: Financial capital of India
          nullable: true

    PaginatedCityResponse:
      type: object
      description: Paginated response containing a list of cities and pagination metadata
      required:
        - content
        - page
        - size
        - totalItems
        - totalPages
      properties:
        content:
          type: array
          description: List of cities on the current page
          items:
            $ref: '#/components/schemas/City'
        page:
          type: integer
          description: Current zero-based page index
          example: 0
        size:
          type: integer
          description: Number of items per page
          example: 10
        totalItems:
          type: integer
          format: int64
          description: Total number of cities for the country
          example: 5
        totalPages:
          type: integer
          description: Total number of pages
          example: 1

    ErrorResponse:
      type: object
      description: Standard error response body
      required:
        - status
        - message
        - timestamp
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 404
        message:
          type: string
          description: Human-readable error description
          example: Country not found with id 999
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred
          example: "2024-01-15T10:30:00Z"